// pwa/src/core/models/elemental-template.ts
export interface ElementalTemplate {
  id?: number;
  characterId: number;
  templateName: string;
  element: number;
  summoningType: number;
  astralSense: boolean;
  longArm: boolean;
  lifeSense: boolean;
  regenerationLevel: number;
  additionalActionsLevel: number;
  resistanceMagic: boolean;
  resistanceTraitDamage: boolean;
  immunityMagic: boolean;
  immunityTraitDamage: boolean;
  resistancesDemonicJson: string;
  resistancesElementalJson: string;
  immunitiesDemonicJson: string;
  immunitiesElementalJson: string;
  causeFear: boolean;
  artifactAnimationLevel: number;
  aura: boolean;
  blinkingInvisibility: boolean;
  elementalShackle: boolean;
  elementalGripLevel: number;
  elementalInferno: boolean;
  elementalGrowth: boolean;
  drowning: boolean;
  areaAttack: boolean;
  flight: boolean;
  frost: boolean;
  ember: boolean;
  criticalImmunity: boolean;
  boilingBlood: boolean;
  fog: boolean;
  smoke: boolean;
  stasis: boolean;
  stoneEatingLevel: number;
  stoneSkinLevel: number;
  mergeWithElement: boolean;
  sinking: boolean;
  wildGrowth: boolean;
  burst: boolean;
  shatteringArmor: boolean;
  modLeP: number;
  modINI: number;
  modRS: number;
  modGS: number;
  modMR: number;
  modAT: number;
  modPA: number;
  modTP: number;
  modAttribute: number;
  modNewTalent: number;
  modTaWZfW: number;
}